refactor: remove abstract partition manager
